Moving set_default_connection() to _implicit_environ.
Moved from datastore.__init__.
