+++ gcloud/datastore/test___init__.py
@@ -15,42 +15,6 @@
 import unittest2
 
 
-class Test_set_default_connection(unittest2.TestCase):
-
-    def setUp(self):
-        from gcloud.datastore._testing import _setup_defaults
-        _setup_defaults(self)
-
-    def tearDown(self):
-        from gcloud.datastore._testing import _tear_down_defaults
-        _tear_down_defaults(self)
-
-    def _callFUT(self, connection=None):
-        from gcloud.datastore import set_default_connection
-        return set_default_connection(connection=connection)
-
-    def test_set_explicit(self):
-        from gcloud.datastore import _implicit_environ
-
-        self.assertEqual(_implicit_environ.get_default_connection(), None)
-        fake_cnxn = object()
-        self._callFUT(connection=fake_cnxn)
-        self.assertEqual(_implicit_environ.get_default_connection(), fake_cnxn)
-
-    def test_set_implicit(self):
-        from gcloud._testing import _Monkey
-        from gcloud import datastore
-        from gcloud.datastore import _implicit_environ
-
-        self.assertEqual(_implicit_environ.get_default_connection(), None)
-
-        fake_cnxn = object()
-        with _Monkey(datastore, get_connection=lambda: fake_cnxn):
-            self._callFUT()
-
-        self.assertEqual(_implicit_environ.get_default_connection(), fake_cnxn)
-
-
 class Test_set_defaults(unittest2.TestCase):
 
     def _callFUT(self, dataset_id=None, connection=None):
